EIPを32bitにする — 最後の16bit配線を外す
ip: u16 はコアに残った最後の16bit配線だった。64KBを超えるコード
(Linuxカーネルは1MB超) はこれが外れないと歩けない。

## 折り返しの持ち主を1箇所にする

widen して終わりではなく、「16bitコードでは64Kで折り返す」という8086の
挙動を**どこが知っているか**を決めるのが本体だった。散らばらせず、
Cpu::ip_mask() / set_ip() / advance_ip() の3つだけが幅を知る。
幅はモードとCSのDビットから導く — regs を最初から u32 で持っていたのと
同じ判断を IP にも適用した形。

- フェッチ (最熱経路) は advance_ip(1)。ANDが1個増えるだけ
- 分岐族 (jmp/call/ret/jcc/loop/grp5) は着地を set_ip に統一。
  ついでに**オペランドサイズ対応** (rel32 / pop32) まで済ませた —
  型を直すだけなら二度手間になる場所
- far call/retf (9A/CB/CA) はリアルモード16bit形のまま。保護モードの
  far call は特権チェックと絡むので、リングを作るときに一緒にやる

## 道連れ

- 割り込み統計 (int_first/int_recent/first_fault) のIPを u32 に
- debug (Step.ip / Stop::*.at / before_exec) を u32 に
- スナップショット v4
- boot.rs の診断にあった sel<<4 直書きの残党2箇所も cpu.lin() に統一

## 検証

- 新テスト2件: **EIPが64Kを越えて歩ける** (0x2_0000 のコードを実行。
  これまでの全テストは偶然64K以下で走っていた) / **リアルモードのIPは
  64Kで折り返す** (8086の環)
- 103テスト全通過、clippy 0
- **ELKSがまたビット同一 (45,669,277命令)** — フェッチ経路を触っても
  リアルモードの挙動はゼロ変化。FreeDOS も dir まで無傷

+++ core/src/lib.rs
@@ -78,13 +78,13 @@ pub struct Machine {
     pub disk: Option<Disk>,
     /// 最初に起きたCPU例外の (ベクタ番号, CS, IP)。
     /// 実OSを動かすと「どこで壊れたか」だけが手がかりになるので控えておく
-    pub first_fault: Option<(u8, u16, u16)>,
+    pub first_fault: Option<(u8, u16, u32)>,
     /// ベクタごとの発生回数。全部数える (周期割り込みで溢れないよう回数だけ)
     pub int_counts: Vec<u32>,
     /// ベクタごとの初出位置 (CS, IP)
-    pub int_first: Vec<(u16, u16)>,
+    pub int_first: Vec<(u16, u32)>,
     /// 直近の割り込み (ベクタ, CS, IP)。**panic直前に何が起きたかはここに出る**
-    pub int_recent: std::collections::VecDeque<(u8, u16, u16)>,
+    pub int_recent: std::collections::VecDeque<(u8, u16, u32)>,
     /// 外から覗くための仕掛け。**機械の状態ではない**のでスナップショットには入れない
     pub dbg: debug::Debug,
     pub halted: bool,
@@ -406,7 +406,7 @@ impl Machine {
         for s in self.cpu.sregs {
             w.u16(s);
         }
-        w.u16(self.cpu.ip);
+        w.u32(self.cpu.ip);
         w.u32(self.cpu.flags);
         // プロテクトモードの状態 (v2)。隠しレジスタを落とすと、復元した瞬間に
         // 全アドレスが嘘になる — セレクタだけでは base を再構成できない
@@ -463,7 +463,7 @@ impl Machine {
         for i in 0..6 {
             m.cpu.sregs[i] = r.u16()?;
         }
-        m.cpu.ip = r.u16()?;
+        m.cpu.ip = r.u32()?;
         m.cpu.flags = r.u32()?;
         m.cpu.cr0 = r.u32()?;
         m.cpu.gdtr_base = r.u32()?;
@@ -619,7 +619,7 @@ impl Machine {
         if self.dbg.on {
             let (cs, ip) = (self.cpu.sregs[cpu::CS], self.cpu.ip);
             // 線形アドレスは隠しレジスタ経由。sel<<4 は保護モードで嘘になる
-            let lin = self.cpu.lin(cpu::CS, ip as u32);
+            let lin = self.cpu.lin(cpu::CS, ip);
             if self.dbg.before_exec(lin, cs, ip) {
                 self.dbg.instr -= 1; // 実行しなかったので数え戻す
                 return;
